+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+- Enhance TabCloseButton component with tabIndex prop for improved accessibility. (`4053939`)
 - Enhance SegmentedTabs component with keyboard navigation improvements. (`6ec88d2`)
 
 ## 1.0.19 - 2026-07-02